refactor Petri.jl and add stochastic tests
add ParamProblem and AbstractProblem
